+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 # Changelog
 ## v21.0.0
 - Update to MC 1.21.
-- Make player armor support transparency.
-- Make player capes support transparency.
+- Make player armor support transparency [#9](https://github.com/Trikzon/transparent/issues/9).
+- Make player capes support transparency [#8](https://github.com/Trikzon/transparent/issues/8).
+- Make end crystal beams support transparency [#20](https://github.com/Trikzon/transparent/issues/20).
 
 ## v20.6.1
 - Correctly tag the files as MC 1.20.6 when publishing